// src/types/company.ts
export interface Company {
  cnpj: string
  name: string
  fantasyName: string
  address: string
  number: string
  neighborhood: string
  cityCode: string
  city: string
  state: string
  zipCode: string
  stateRegistration: string
  taxRegime: string // 1: Simples Nacional, 2: Lucro Presumido, 3: Lucro Real
}